+++ Source/EventFlow.MsSql.Tests/UnitTests/ReadModels/ReadModelSqlGeneratorTests.cs
@@ -1,31 +1,33 @@
 ﻿// The MIT License (MIT)
-//
+// 
 // Copyright (c) 2015 Rasmus Mikkelsen
+// Copyright (c) 2015 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
+// 
 using System.ComponentModel.DataAnnotations.Schema;
-using EventFlow.MsSql.Tests.ReadModels;
+using EventFlow.MsSql.Tests.IntegrationTests.ReadStores.ReadModels;
 using EventFlow.ReadStores.MsSql;
 using EventFlow.TestHelpers;
 using FluentAssertions;
 using NUnit.Framework;
+#pragma warning disable 618
 
 namespace EventFlow.MsSql.Tests.UnitTests.ReadModels
 {
@@ -38,11 +40,11 @@ namespace EventFlow.MsSql.Tests.UnitTests.ReadModels
         public void CreateInsertSql_ProducesCorrectSql()
         {
             // Act
-            var sql = Sut.CreateInsertSql<MsSqlTestAggregateReadModel>();
+            var sql = Sut.CreateInsertSql<MsSqlThingyReadModel>();
 
             // Assert
             sql.Should().Be(
-                "INSERT INTO [ReadModel-TestAggregate] " +
+                "INSERT INTO [ReadModel-ThingyAggregate] " +
                 "(AggregateId, CreateTime, DomainErrorAfterFirstReceived, LastAggregateSequenceNumber, PingsReceived, UpdatedTime) " +
                 "VALUES " +
                 "(@AggregateId, @CreateTime, @DomainErrorAfterFirstReceived, @LastAggregateSequenceNumber, @PingsReceived, @UpdatedTime)");
@@ -52,11 +54,11 @@ namespace EventFlow.MsSql.Tests.UnitTests.ReadModels
         public void CreateUpdateSql_ProducesCorrectSql()
         {
             // Act
-            var sql = Sut.CreateUpdateSql<MsSqlTestAggregateReadModel>();
+            var sql = Sut.CreateUpdateSql<MsSqlThingyReadModel>();
 
             // Assert
             sql.Should().Be(
-                "UPDATE [ReadModel-TestAggregate] SET " +
+                "UPDATE [ReadModel-ThingyAggregate] SET " +
                 "CreateTime = @CreateTime, DomainErrorAfterFirstReceived = @DomainErrorAfterFirstReceived, " +
                 "LastAggregateSequenceNumber = @LastAggregateSequenceNumber, " +
                 "PingsReceived = @PingsReceived, UpdatedTime = @UpdatedTime " +
@@ -67,10 +69,10 @@ namespace EventFlow.MsSql.Tests.UnitTests.ReadModels
         public void CreateSelectSql_ProducesCorrectSql()
         {
             // Act
-            var sql = Sut.CreateSelectSql<MsSqlTestAggregateReadModel>();
+            var sql = Sut.CreateSelectSql<MsSqlThingyReadModel>();
 
             // Assert
-            sql.Should().Be("SELECT * FROM [ReadModel-TestAggregate] WHERE AggregateId = @AggregateId");
+            sql.Should().Be("SELECT * FROM [ReadModel-ThingyAggregate] WHERE AggregateId = @EventFlowReadModelId");
         }
 
         [Test]