+++ Source/EventFlow.ReadStores.Elasticsearch/ElasticsearchReadModelStore.cs
@@ -1,25 +1,26 @@
 ﻿// The MIT License (MIT)
-//
+// 
 // Copyright (c) 2015 Rasmus Mikkelsen
+// Copyright (c) 2015 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
+// 
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -66,11 +67,11 @@ namespace EventFlow.ReadStores.Elasticsearch
 
             if (!getResponse.IsValid || !getResponse.Found)
             {
-                return ReadModelEnvelope<TReadModel>.Empty;
+                return ReadModelEnvelope<TReadModel>.Empty(id);
             }
 
             var version = long.Parse(getResponse.Version);
-            return ReadModelEnvelope<TReadModel>.With(getResponse.Source, version);
+            return ReadModelEnvelope<TReadModel>.With(id, getResponse.Source, version);
         }
 
         public Task DeleteAllAsync(
@@ -108,8 +109,8 @@ namespace EventFlow.ReadStores.Elasticsearch
                     .ConfigureAwait(false);
 
                 var readModelEnvelope = response.Found
-                    ? ReadModelEnvelope<TReadModel>.With(response.Source, long.Parse(response.Version))
-                    : ReadModelEnvelope<TReadModel>.Empty;
+                    ? ReadModelEnvelope<TReadModel>.With(readModelUpdate.ReadModelId, response.Source, long.Parse(response.Version))
+                    : ReadModelEnvelope<TReadModel>.Empty(readModelUpdate.ReadModelId);
 
                 readModelEnvelope = await updateReadModel(readModelContext, readModelUpdate.DomainEvents, readModelEnvelope, cancellationToken).ConfigureAwait(false);