+++ Source/EventFlow.Tests/IntegrationTests/ResolverTests.cs
@@ -1,29 +1,30 @@
 ﻿// The MIT License (MIT)
-//
+// 
 // Copyright (c) 2015 Rasmus Mikkelsen
+// Copyright (c) 2015 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
+// 
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Extensions;
-using EventFlow.TestHelpers.Aggregates.Test;
+using EventFlow.TestHelpers.Aggregates;
 using FluentAssertions;
 using NUnit.Framework;
 
@@ -33,11 +34,11 @@ namespace EventFlow.Tests.IntegrationTests
     {
         public class Service { }
 
-        public class ServiceDependentAggregate : AggregateRoot<ServiceDependentAggregate, TestId>
+        public class ServiceDependentAggregate : AggregateRoot<ServiceDependentAggregate, ThingyId>
         {
             public Service Service { get; }
 
-            public ServiceDependentAggregate(TestId id, Service service) : base(id)
+            public ServiceDependentAggregate(ThingyId id, Service service) : base(id)
             {
                 Service = service;
             }
@@ -56,7 +57,7 @@ namespace EventFlow.Tests.IntegrationTests
                 var aggregateFactory = resolver.Resolve<IAggregateFactory>();
 
                 // Act
-                var serviceDependentAggregate = await aggregateFactory.CreateNewAggregateAsync<ServiceDependentAggregate, TestId>(TestId.New).ConfigureAwait(false);
+                var serviceDependentAggregate = await aggregateFactory.CreateNewAggregateAsync<ServiceDependentAggregate, ThingyId>(ThingyId.New).ConfigureAwait(false);
 
                 // Assert
                 serviceDependentAggregate.Service.Should()