+++ Source/EventFlow.Tests/UnitTests/Commands/DistinctCommandTests.cs
@@ -1,30 +1,31 @@
 ﻿// The MIT License (MIT)
-//
+// 
 // Copyright (c) 2015 Rasmus Mikkelsen
+// Copyright (c) 2015 eBay Software Foundation
 // https://github.com/rasmus/EventFlow
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of
 // this software and associated documentation files (the "Software"), to deal in
 // the Software without restriction, including without limitation the rights to
 // use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
 // the Software, and to permit persons to whom the Software is furnished to do so,
 // subject to the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
 // FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
+// 
 using System;
 using System.Collections.Generic;
 using EventFlow.Commands;
 using EventFlow.Extensions;
-using EventFlow.TestHelpers.Aggregates.Test;
+using EventFlow.TestHelpers.Aggregates;
 using FluentAssertions;
 using NUnit.Framework;
 
@@ -33,12 +34,12 @@ namespace EventFlow.Tests.UnitTests.Commands
     [Timeout(10000)]
     public class DistinctCommandTests
     {
-        public class MyDistinctCommand : DistinctCommand<TestAggregate, TestId>
+        public class MyDistinctCommand : DistinctCommand<ThingyAggregate, ThingyId>
         {
             public int MagicNumber { get; }
 
             public MyDistinctCommand(
-                TestId aggregateId,
+                ThingyId aggregateId,
                 int magicNumber) : base(aggregateId)
             {
                 MagicNumber = magicNumber;
@@ -51,13 +52,13 @@ namespace EventFlow.Tests.UnitTests.Commands
             }
         }
 
-        [TestCase("test-4b1e7b48-18f1-4215-91d9-903cffdab3d8", 1, "command-ed8fe9cc-0b14-5e3a-b2fd-8b6b29cdffbb")]
-        [TestCase("test-4b1e7b48-18f1-4215-91d9-903cffdab3d8", 2, "command-9e71d2cd-18c2-517c-b58f-6e4faf3b9b04")]
-        [TestCase("test-6a2a04bd-bbc8-44ac-80ac-b0ca56897bc0", 2, "command-db2711e2-6947-5fe4-87c6-13ca92438881")]
+        [TestCase("thingy-4b1e7b48-18f1-4215-91d9-903cffdab3d8", 1, "command-40867fe4-94da-59e4-9bb5-e532f6565751")]
+        [TestCase("thingy-4b1e7b48-18f1-4215-91d9-903cffdab3d8", 2, "command-df0e238e-676b-500e-b962-76fcef97768a")]
+        [TestCase("thingy-6a2a04bd-bbc8-44ac-80ac-b0ca56897bc0", 2, "command-0455a861-bc9e-56c5-b7b9-5c14671db8b2")]
         public void Arguments(string aggregateId, int magicNumber, string expectedSouceId)
         {
             // Arrange
-            var testId = TestId.With(aggregateId);
+            var testId = ThingyId.With(aggregateId);
             var command = new MyDistinctCommand(testId, magicNumber);
 
             // Act